+++ .eslintrc.js
@@ -1,6 +1,7 @@
 module.exports = {
   extends: ["@lad-tech/eslint-config"],
   rules: {
+    "react/react-in-jsx-scope": 0,
     "prettier/prettier": [
       "error",
       {
