+++ src/shared/ui/index.tsx
@@ -1,0 +1,1 @@
+export * from "react-native-paper";
